Split PushNotification module into two
As ita requested, we should use the name wor-push-notifications-aws instead of wor-push_notifications-aws.
This change requires to add an additional module and because of that, we have to change several conventions used in both modules and specs.

It's good to know that, it was also proved with a dummy app and worked fine.